added resume capability

+++ src/test/test.py
@@ -3,29 +3,22 @@ import os
 import argparse
 import multiprocessing
 import sys
-import json
 
 
 def main(args):
-    global success, compile_error, test_error
+    global test_results
 
     lines = []
     with open(f'data/{args.project_name}/unique_methods_{args.model_name}_selected_bugs.jsonl') as f:
         lines = f.readlines()
 
-    manager = multiprocessing.Manager()
-    success = manager.list()
-    compile_error = manager.list()
-    test_error = manager.list()
+    test_results = open(f'data/{args.project_name}/test_results_{args.model_name}.txt', 'a')
+
+    os.makedirs(f'test_results/{args.project_name}', exist_ok=True)
 
     pool = multiprocessing.Pool(args.num_workers)
     for i, _ in enumerate(pool.imap_unordered(process_instance, lines), 1):
         sys.stderr.write('\rpercentage of source code files completed: {0:%}'.format(i/len(lines)))
-
-    stats = {'success': list(success), 'compile_error': list(compile_error), 'test_error': list(test_error)}
-
-    with open(f'data/{args.project_name}/test_results_{args.model_name}.json', 'w') as f:
-        json.dump(stats, f)
 
 
 def process_instance(l):
@@ -39,10 +32,14 @@ def process_instance(l):
 
     for bug_id in dct['selected_bugs']:
         project = dct['project']
+
         os.makedirs(f'temp_project_{index}', exist_ok=True)
 
+        if os.path.exists(f'test_results/{project}/{project}.{index}.{bug_id}.{args.model_name}.build.log'):
+            continue
+
         os.system(f'cp -r projects/{project} temp_project_{index}/')
-                
+
         buggy_method = dct[f'buggy_method{bug_id}']
 
         file_path = dct['file_path']
@@ -72,12 +69,17 @@ def process_instance(l):
         data = ''.join(data)
 
         if 'BUILD SUCCESS' in data:
-            success.append(f'{index}-{bug_id}')
+            test_results.write(f'{index}-{bug_id}-0\n')
+            test_results.flush()
         elif 'BUILD FAILURE' in data and 'COMPILATION ERROR' in data:
-            compile_error.append(f'{index}-{bug_id}')
+            test_results.write(f'{index}-{bug_id}-1\n')
+            test_results.flush()
         elif 'BUILD FAILURE' in data:
-            test_error.append(f'{index}-{bug_id}')
+            test_results.write(f'{index}-{bug_id}-2\n')
+            test_results.flush()
         
+        os.system(f'echo > ../../test_results/{project}/{project}.{index}.{bug_id}.{args.model_name}.build.log')
+
         os.chdir('../../')
         os.system(f'rm -rf temp_project_{index}/{project}')
